+++ src/comp/List.tsx
@@ -1,6 +1,7 @@
 import { useEffect, useRef, useState } from "react";
 import { type ActionData } from "swiper-action";
-import { searchWord } from "../service/search.service";
+import { env } from "../env/client.mjs";
+import { searchWord } from "../server/service/search.service";
 import Error from "../sites/Error";
 import { type ListElement } from "../types/types";
 import {
@@ -11,7 +12,6 @@ import {
   resetIcon,
 } from "../utils/icons";
 import ListItem from "./ListItem";
-import { env } from "../env/client.mjs";
 
 type Props = {
   words: ListElement[];
