+++ src/server/repository/WotdSupabaseRepository.ts
@@ -2,8 +2,8 @@ import type { Word } from "@prisma/client";
 import { type Tag, type VocabularyWord, type WOTD } from "../../types/types";
 import { addIcons } from "../../utils/helper";
 import { prisma } from "../db";
+import { getTodayMorning } from "../service/getDate.service";
 import { type WOTDRepository } from "./WOTDRepository";
-import { getTodayMorning } from "../../service/getDate.service";
 
 export class WOTDSupabaseRepository implements WOTDRepository {
   getToday = async () => {
