+++ src/server/service/parseCache.service.ts
@@ -1,7 +1,7 @@
 /* eslint-disable jsdoc/require-jsdoc */
-import type { ListElement, Settings } from "../types/types";
-import { isVocabSettings } from "../utils/guards/other";
-import { isListElement } from "../utils/guards/words";
+import type { ListElement, Settings } from "../../types/types";
+import { isVocabSettings } from "../../utils/guards/other";
+import { isListElement } from "../../utils/guards/words";
 
 export function parseListElements(input: string): ListElement[] {
   const parsed = JSON.parse(input) as unknown;
